requirements.txt
streamlit
rdkit-pypi
py3Dmol
scikit-learn
joblib
numpy
pandas
graphviz
Pillow
requests
matplotlib
